RAASU/v1.0.1/requirements.txt
appdirs==1.4.4
argcomplete==3.6.3
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
docutils==0.22.4
et_xmlfile==2.0.0
filetype==1.2.0
idna==3.11
importlib_resources==6.5.2
Kivy==2.3.1
Kivy-Garden==0.1.5
openpyxl==3.1.5
pillow==12.1.1
Pygments==2.19.2
pyserial==3.5
python-barcode==0.16.1
python-escpos==3.1
pyusb==1.3.1
PyYAML==6.0.3
qrcode==8.2
requests==2.32.5
setuptools==82.0.0
six==1.17.0
svgwrite==1.4.3
Tree==0.2.4
urllib3==2.6.3
